+++ tests/integration/cypress/e2e/main.cy.ts
@@ -3,7 +3,7 @@ import { faker } from '@faker-js/faker'
 describe('Visual Regression Example', () => {
   it('should display the home page correctly', () => {
     cy.visit('./cypress/web/01.html')
-    cy.get('H1').contains('Hello, World').should('exist')
+    cy.get('div').contains('Hello, World').should('exist')
     cy.compareSnapshot('home')
   })
 
@@ -20,20 +20,20 @@ describe('Visual Regression Example', () => {
 
   it('should display the register page correctly', () => {
     cy.visit('./cypress/web/02.html')
-    cy.get('H1').contains('Register').should('exist')
+    cy.get('div').contains('Register').should('exist')
     cy.compareSnapshot('register')
   })
 
   it('should display the login page correctly', () => {
     cy.visit('./cypress/web/03.html')
-    cy.get('H1').contains('Login').should('exist')
+    cy.get('div').contains('Login').should('exist')
     cy.compareSnapshot('login', 0.0)
     cy.compareSnapshot('login', 0.1)
   })
 
   it('should display the component correctly', () => {
     cy.visit('./cypress/web/03.html')
-    cy.get('H1').contains('Login').should('exist')
+    cy.get('div').contains('Login').should('exist')
     cy.get('form')
       .compareSnapshot('login-form')
       .should((result) => {
@@ -79,10 +79,10 @@ describe('Visual Regression Example', () => {
     })
     if (Cypress.env('visualRegressionType') === 'base') {
       cy.visit('./cypress/web/04.html')
-      cy.get('H1').contains('bar').should('exist')
+      cy.get('div').contains('bar').should('exist')
     } else {
       cy.visit('./cypress/web/05.html')
-      cy.get('H1').contains('none').should('exist')
+      cy.get('div').contains('none').should('exist')
     }
     cy.compareSnapshot('bar')
   })
@@ -91,26 +91,26 @@ describe('Visual Regression Example', () => {
     if (Cypress.env('visualRegressionType') === 'base') {
       cy.visit('./cypress/web/04.html')
       cy.compareSnapshot('foo')
-      cy.get('H1').compareSnapshot('h1')
+      cy.get('div').compareSnapshot('bar')
     } else {
       cy.visit('./cypress/web/05.html')
-      cy.get('H1').contains('none').should('exist')
-      cy.compareSnapshot('foo', 0.02).should((result) => {
+      cy.get('div').contains('none').should('exist')
+      cy.compareSnapshot('foo', 0.2).should((result) => {
         expect(result.error).to.be.undefined
-        expect(result.percentage).to.be.below(0.02)
+        expect(result.percentage).to.be.below(0.2)
       })
       cy.compareSnapshot('foo', { errorThreshold: 0.01, failSilently: true }).should((result) => {
         expect(result.percentage).to.be.above(0.01)
         expect(result.error).to.exist
       })
-      cy.get('H1')
-        .compareSnapshot('h1', 0.085)
+      cy.get('div')
+        .compareSnapshot('bar', 0.085)
         .should((result) => {
           expect(result.error).to.be.undefined
           expect(result.percentage).to.be.below(0.085)
         })
-      cy.get('H1')
-        .compareSnapshot('h1', { errorThreshold: 0.01, failSilently: true })
+      cy.get('div')
+        .compareSnapshot('bar', { errorThreshold: 0.01, failSilently: true })
         .should((result) => {
           expect(result.percentage).to.be.above(0.01)
           expect(result.images.actual).to.exist
@@ -127,7 +127,7 @@ describe('Visual Regression Example', () => {
     } else {
       cy.visit('./cypress/web/07.html')
     }
-    cy.get('H1').contains('Color').should('exist')
+    cy.get('div').contains('Color').should('exist')
     cy.compareSnapshot('baz', 0.029)
     cy.compareSnapshot('baz', { errorThreshold: 0.02, failSilently: true })
     cy.compareSnapshot('baz', { failSilently: true })
@@ -142,11 +142,11 @@ describe('Visual Regression Example', () => {
     })
     if (Cypress.env('visualRegressionType') === 'base') {
       cy.visit('./cypress/web/07.html')
-      cy.get('H1').contains('Color').should('exist')
+      cy.get('div').contains('Color').should('exist')
       cy.compareSnapshot('bar-07', 0.1)
     } else {
       cy.visit('./cypress/web/08.html')
-      cy.get('H1').contains('Color').should('exist')
+      cy.get('div').contains('Color').should('exist')
       cy.compareSnapshot('bar-07', 0.1).then((result) => {
         expect(result.error).to.exist
       })
@@ -175,17 +175,17 @@ describe('Visual Regression Example', () => {
     () => {
       if (Cypress.env('visualRegressionType') === 'base') {
         cy.visit('./cypress/web/04.html')
-        cy.get('H1').contains('bar').should('exist')
+        cy.get('div').contains('bar').should('exist')
         cy.compareSnapshot('foo')
-        cy.get('H1').compareSnapshot('h1')
+        cy.get('div').compareSnapshot('bar')
       } else {
         cy.visit('./cypress/web/05.html')
-        cy.get('H1').contains('none').should('exist')
+        cy.get('div').contains('none').should('exist')
         cy.compareSnapshot('foo', 0.01).should((result) => {
           expect(result.error).to.exist
         })
-        cy.get('H1')
-          .compareSnapshot('h1', 0.02)
+        cy.get('div')
+          .compareSnapshot('bar', 0.02)
           .should((result) => {
             expect(result.error).to.exist
           })
@@ -193,7 +193,7 @@ describe('Visual Regression Example', () => {
     }
   )
 
-  it('should log command options to Cypress.log', () => {
+  it.only('should log command options to Cypress.log', () => {
     cy.on('fail', (error) => {
       if (error.message.includes("The 'random' image is different.")) {
         return
@@ -236,7 +236,7 @@ describe('Visual Regression Example', () => {
       }
     })
     cy.visit('./cypress/web/random.html')
-    cy.get('H1').should('be.visible')
+    cy.get('div').should('be.visible')
     cy.compareSnapshot('random')
   })